Split cuda_utils in to host and device parts. Include directive cleanup.
