+++ dali/test/dali_cuda_finalize_test.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2020-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,8 +17,7 @@
 
 #include <cuda_runtime.h>
 #include <gtest/gtest.h>
-
-#include "dali/core/cuda_utils.h"
+#include "dali/core/cuda_error.h"
 
 namespace dali {
